requirements.txt
feedparser==6.0.11
newsapi-python==0.2.7
requests==2.31.0
groq==0.9.0
python-dotenv==1.0.0
google-auth==2.23.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0